+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixed
+- Fixed a bug where workload with subgroups would not consider additional tasks above minAvailable
+
 ## [v0.8.3] - 20250-8-31
 
 ### Removed